perf: fix project rebuild even if no files changed
Fix the project being rebuilt even if no files changed. This was
introduced by #22, and the mistake was that `migrations/` and `preset/`
are directories at the project root, but `build.rs` is in `lyra/` and
the cargo emits in it expects the current working directory to be
`lyra/` and not project root. As the directory doesn't exist in the
specified location, it rebuilds (which is [the documented
behaviour](https://doc.rust-lang.org/cargo/faq.html#why-is-cargo-rebuilding-my-code)).