build: centralise workspace configuration and deps (#100)
* chore: update Cargo.lock, upgrade cargo deps

* fix: pin devenv to v1.6.1 to avoid v1.7.0 for now

devenv v1.7.0 is released, but it includes a bug that broke the rust
toolchain creation (https://github.com/cachix/devenv/pull/1983), so it
is not usable for rust development at the moment. This commits pin
devenv to v1.6.1 in the meantime.

* chore: update flake.lock

* build: centralise workspace configuration and deps

Centralise workspace configuration (package metadata, lint settings for
rustc and clippy) and dependencies in the root `Cargo.toml` file, while
making each sub-crates' `Cargo.toml` inherit relevant properties from
it.